# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VIDULA API",
        "description": "REST API — Sanctum token auth",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/auth/otp/send": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send OTP code via Email or SMS",
                "operationId": "21e5f5606eae374de63549b0be74ed69",
                "responses": {
                    "200": {
                        "description": "OTP sent"
                    }
                }
            }
        },
        "/api/auth/otp/verify": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify OTP code and retrieve access token",
                "operationId": "7606bb8c6d5bf76303a57a94b0969b86",
                "responses": {
                    "200": {
                        "description": "Returns Sanctum access token"
                    },
                    "401": {
                        "description": "Invalid OTP or user"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout and revoke current token",
                "operationId": "654d4a72d1cb086ee8044f23ede24ed4",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/{provider}/callback": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login with Social Provider Token",
                "operationId": "5b796df7749fcdda7cf94b898c552a8d",
                "responses": {
                    "200": {
                        "description": "Returns Sanctum access token"
                    }
                }
            }
        },
        "/api/products/admin": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List products (paginated)",
                "operationId": "fc0ad78759ec7503da4d76ed24ca571c",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductListItem"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create product",
                "operationId": "227f3c48c31d354af7af91add543d06b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/admin/{uuid}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get single product details",
                "operationId": "a7070ffe74a1db0ab5860a6556821105",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product details",
                "operationId": "2602f8cce95f8b6dd01a8fef5cbc80ec",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Soft delete product",
                "operationId": "af63fd014cbb8c600efd322cc83c0461",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/admin/bulk-delete": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Bulk soft delete products",
                "operationId": "16d51047ccc160ba4d16b5f018b2e52f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Products deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/admin/{uuid}/restore": {
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Restore soft deleted product",
                "operationId": "a6569f5e7f5187d55078f0b8148681dd",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product restored successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CreateProductDTO": {
                "required": [
                    "userId",
                    "type",
                    "title",
                    "slug",
                    "price",
                    "currency",
                    "level",
                    "language"
                ],
                "properties": {
                    "userId": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "number"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "level": {
                        "type": "string"
                    },
                    "language": {
                        "type": "string"
                    },
                    "thumbnail": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateProductDTO": {
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "number"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "level": {
                        "type": "string"
                    },
                    "language": {
                        "type": "string"
                    },
                    "thumbnail": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ProductListItem": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "price": {
                        "type": "number"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "thumbnail": {
                        "type": "string",
                        "nullable": true
                    },
                    "level": {
                        "type": "string"
                    },
                    "language": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}